# Fix account balance file ingestion performance issue