--- a/hedera-mirror-importer/src/main/resources/db/migration/v1/V1.51.0__drop_account_balance_index.sql
+++ b/hedera-mirror-importer/src/main/resources/db/migration/v1/V1.51.0__drop_account_balance_index.sql
@@ -1,0 +1,5 @@
+-------------------
+-- Drop the index account_balance__account_timestamp
+-------------------
+
+drop index if exists account_balance__account_timestamp;