--- a/hedera-mirror-importer/src/main/resources/db/migration/v2/V2.0.3__index_init.sql
+++ b/hedera-mirror-importer/src/main/resources/db/migration/v2/V2.0.3__index_init.sql
@@ -12,8 +12,6 @@ create index if not exists assessed_custom_fee__consensus_timestamp
 -- account_balance
 alter table account_balance
     add constraint account_balance__pk primary key (consensus_timestamp, account_id);
-create index if not exists account_balance__account_timestamp
-    on account_balance (account_id desc, consensus_timestamp desc);
 
 -- account_balance_file
 alter table account_balance_file