--- a/hedera-mirror-rest/accounts.js
+++ b/hedera-mirror-rest/accounts.js
@@ -73,7 +73,6 @@ const entityAndBalanceFields = [
   'balance account_balance',
 ].join(',');
 const latestBalanceFilter = 'ab.consensus_timestamp = (select max(consensus_timestamp) from account_balance)';
-const latestBalanceNullableFilter = `(${latestBalanceFilter} or ab.consensus_timestamp is null)`;
 
 /**
  * Gets the query for entity fields with hbar balance info for the full outer join case.
@@ -151,8 +150,8 @@ const getEntityBalanceQuery = (
   order,
   pubKeyQuery
 ) => {
+  const balanceJoinConditions = `ab.account_id = e.id and ${latestBalanceFilter}`;
   const balanceWhereConditions = [balanceQuery.query];
-  let balanceTimestampFilter = latestBalanceFilter;
   let entityIdField = 'id'; // use 'id' from account / contract for inner and left outer joins
   let joinType = '';
   let params;
@@ -165,8 +164,6 @@ const getEntityBalanceQuery = (
     joinType = 'inner';
     params = [entityAccountQuery.params, pubKeyQuery.params, balanceQuery.params];
   } else if (pubKeyQuery.query) {
-    // allow null for account_balance consensus_timestamp for left outer join
-    balanceTimestampFilter = latestBalanceNullableFilter;
     joinType = 'left outer';
     params = [entityAccountQuery.params, pubKeyQuery.params];
   } else if (balanceQuery.query) {
@@ -177,7 +174,6 @@ const getEntityBalanceQuery = (
     params = [balanceQuery.params, balanceAccountQuery.params];
   }
 
-  balanceWhereConditions.push(balanceTimestampFilter);
   const whereCondition = [
     // no entity id filter needed for account / contract for right outer join
     joinType !== 'right outer' ? entityAccountQuery.query : '',
@@ -193,7 +189,7 @@ const getEntityBalanceQuery = (
       select ${entityIdField} id,${entityAndBalanceFields}
       from (${accountContractQuery}) e
       ${joinType} join account_balance ab
-        on ab.account_id = e.id
+        on ${balanceJoinConditions}
       ${whereClause}
       order by ${entityIdField} ${order}
       ${limitQuery}